+++ lib/servus/support/response.rb
@@ -2,15 +2,28 @@
 
 module Servus
   module Support
+    # Response class for service results
     class Response
-      attr_reader :data, :error
+      # [Object] The data returned by the service
+      attr_reader :data
 
+      # [Servus::Support::Errors::ServiceError] The error returned by the service
+      attr_reader :error
+
+      # Initializes a new response
+      #
+      # @param success [Boolean] Whether the response was successful
+      # @param data [Object] The data returned by the service
+      # @param error [Servus::Support::Errors::ServiceError] The error returned by the service
       def initialize(success, data, error)
         @success = success
         @data = data
         @error = error
       end
 
+      # Returns whether the response was successful
+      #
+      # @return [Boolean] Whether the response was successful
       def success?
         @success
       end