+++ spec/servus_spec.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-require "spec_helper"
+require 'spec_helper'
 
 RSpec.describe Servus do
-  it "has a version number" do
+  it 'has a version number' do
     expect(Servus::VERSION).not_to be nil
   end
 end