+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [Unreleased]
 
+## [0.1.1] - 2025-08-20
+
+- Added: Added `rescue_from` method to `Servus::Base` to rescue from standard errors and use custom error types.
+- Added: Added `run_service` and `render_service_object_error` helpers to `Servus::Helpers::ControllerHelpers`.
+- Fixed: All rubocop warnings.
+
 ## [0.1.0] - 2025-04-28
 
 - Initial release